fix(tests): gcc-14 conversion-clean slot-claim host build

+++ lichen/subsys/lichen/coap/coap_slot_coord.c
@@ -261,7 +261,10 @@ bool lichen_slot_coord_validate_interleaved(const uint8_t *slots,
 	}
 
 	for (uint8_t i = 0; i < slot_count; i++) {
-		uint8_t expected = ordinal + i * gateway_count;
+		/* ordinal + i * gateway_count <= 255 by construction
+		 * (slot_count slots, each < gateway_count apart), but int
+		 * promotion trips -Werror=conversion on strict host builds. */
+		uint8_t expected = (uint8_t)(ordinal + i * gateway_count);
 		if (slots[i] != expected) {
 			return false;
 		}
